add common symlinked types
