+++ packages/docusaurus-plugin-redoc/src/types/common.d.ts
@@ -1,0 +1,1 @@
+../../../docusaurus-theme-redoc/src/types/common.d.ts
